Do not fail uninstall_preflight if files missing
Fixes homebrew/homebrew-cask#67983

+++ Casks/adobe-air-sdk.rb
@@ -50,7 +50,7 @@ cask 'adobe-air-sdk' do
   end
 
   uninstall_postflight do
-    FileUtils.rm("#{HOMEBREW_PREFIX}/share/adobe-air-sdk")
+    FileUtils.rm("#{HOMEBREW_PREFIX}/share/adobe-air-sdk") if Pathname.new("#{HOMEBREW_PREFIX}/share/adobe-air-sdk").exist?
   end
 
   caveats <<~EOS
